backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
spacy>=3.7.0
networkx>=3.1
scikit-learn>=1.3.0
numpy>=1.24.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
